+++ templates/dashboard/base.html
@@ -412,6 +412,11 @@
 
                   <ul>
                       {% if perms.reports.view_sale_reports %}
+                        <li class="{% block menu_report_orders_class %}{% endblock %}">
+                          <a href="{% url 'dashboard:orders_list' %}"><i class='icon-cash3'></i>Orders</a>
+                        </li>
+                      {% endif %}
+                      {% if perms.reports.view_sale_reports %}
                         <li class="{% block menu_report_class %}{% endblock %}">
                           <a href="{% url 'dashboard:sales_list' %}"><i class='icon-cash3'></i>Sales</a>
                         </li>
