+++ templates/dashboard/reports/orders/ajax/discount_paginate.html
@@ -1,0 +1,15 @@
+{% load staticfiles %}
+{% load prices_i18n %}
+{% load humanize %}
+{% if sales_by_category %}
+   {% for sales in sales_by_category %}
+  <tr class='clickable-row'>
+    <td class="td">{{sales.product_name}}</td>
+    <td class="td">{{sales.discount}}</td>
+    <td class="td">{{sales.quantity__sum}}</td>
+      <td class="td"><span class="pull-right">{{sales.total_cost__sum}}</span></td>
+  </tr>
+{% endfor %}
+ {% else %}
+ <tr><td colspan="3"><h2 class="text-center">No Category Sales Available</h2></td></tr>
+{% endif %}
