+++ templates/dashboard/reports/orders/ajax/user_paginate.html
@@ -1,0 +1,14 @@
+{% load staticfiles %}
+{% load prices_i18n %}
+{% load humanize %}
+{% if sales_by_category %}
+   {% for sales in sales_by_category %}
+  <tr class='clickable-row'>
+    <td class="td">{{sales.user__name}}</td>
+    <td class="td">{{sales.solditems__quantity__sum}}</td>
+      <td class="td"><span class="pull-right">{{sales.total_net__sum}}</span></td>
+  </tr>
+{% endfor %}
+ {% else %}
+ <tr><td colspan="3"><h2 class="text-center">No User Sales Available</h2></td></tr>
+{% endif %}
