+++ templates/dashboard/reports/orders/charts/pdf/pdf.html
@@ -1,0 +1,132 @@
+ {% load staticfiles %}
+ <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+ <html>
+    <link href='https://fonts.googleapis.com/css?family=Nunito:400,300' rel='stylesheet' type='text/css'>
+     <head>
+         <title>Sales Report</title>
+         <style type="text/css">
+                          body {sans-serif; color: #384047;}
+             .header {font-family: 'Verdana';font-size: 15px;font-weight: 100;text-align: center;color: #007cae;}
+             .title {font-size: 16px;font-weight: 100;}
+             .header p{color:#384047;}
+
+             .img-circle{border-radius: 50%;}
+             table {ax-width: 960px; margin: 10px auto;}
+            tr{border-bottom:0.5px solid #ddd;border-right:0.5px solid #ddd;
+                border-left:0.5px solid #ddd;
+            }
+            td{border:none;}
+            th {
+                  font-weight: bold;
+                  background:#2196F3;
+                  color: #FFF;
+                  border:none;
+                  font-size:11px;
+                  vertical-align:middle;
+                }
+             .td{text-align: center;font-size:11px;}
+             tr{padding:5px;}
+             .img{width:50px;height:50px;}
+         </style>
+     </head>
+     <body>
+         <div class='wrapper'>
+             <div class='header'>
+                 <p class='title' style="margin-top:2px;margin-bottom:0px;font-size:13px;">SALES SUMMARY REPORT <br>
+                   <span style="font-size: 0.8em;">{{today| date:'jS F Y'}}</span>
+                 </p>
+             </div>
+         <div>
+             <img alt="embedded" src="data:image/png;base64,{{image}}"/>
+             <div class="data-table-container">
+                    <h4 style="text-align:center;"> Sales Summary </h4>
+                  <table class="table bordered highlight responsive data-table" style="border:0.5px solid #ddd;">
+                    <thead>
+                      <tr>
+                        <th>
+                        </th>
+                        <th class="right-align">
+                          Sales
+                        </th>
+                      </tr>
+                    </thead>
+                    <tbody>
+                        <tr data-action-go="">
+                          <td>
+                            <span style="font-weight: bold;">Gross Sales</span>
+                          </td>
+                          <td class="right-align" style="text-align:center;">
+                            <div style="font-weight: bold;"><span style="font-size: 0.7em;">Ksh.</span>
+                              {{ date_gross_sales }}
+                            </div>
+                          </td>
+                        </tr>
+                        <tr>
+                          <td>Discounts</td>
+                          <td class="right-align" style="text-align:center;">
+                              <div style="font-weight: bold;"><span style="font-size: 0.7em;">Ksh.</span>
+                              {{ date_total_discount }}
+                            </div>
+                          </td>
+                        </tr>
+                        <tr>
+                          <td>Tax</td>
+                          <td class="right-align" style="text-align:center;">
+                              <div style="font-weight: bold;"><span style="font-size: 0.7em;">Ksh.</span>
+                              {{ date_total_tax }}
+                            </div>
+                          </td>
+                        </tr>
+                        <tr data-action-go="">
+                          <td>
+                            <span style="font-weight: bold;">Net Sales</span>
+                          </td>
+                          <td class="right-align" style="text-align:center;">
+                            <div style="font-weight: bold;"><span style="font-size: 0.7em;">Ksh.</span>
+                              {{ date_total_sales }}
+                            </div>
+                          </td>
+                        </tr>
+                    </tbody>
+                  </table>
+                </div>
+
+                <div class="data-table-container">
+                    <h4 style="text-align:center;"> Cashiers Present </h4>
+                  <table class="table bordered highlight responsive data-table" style="border:0.5px solid #ddd;">
+                      <thead>
+                      <tr>
+                        <th>
+                          Cashier Name
+                        </th>
+                        <th class="td">
+                          Terminal
+                        </th>
+                        <th> Total Sales</th>
+                      </tr>
+                    </thead>
+                    <tbody>
+                      {% for user in users %}
+                        <tr data-action-go="">
+                          <td class="td">
+                            {{ user.user__name }} - {{user.user__email}}
+                          </td>
+                          <td class="td">
+                            {{ user.terminal }}
+                          </td>
+                          <td class="td">
+                            <div><span style="font-size: 0.7em;">Ksh. </span>{{ user.total_net__sum }}</div>
+                          </td>
+                        </tr>
+                      {% endfor %}
+                    </tbody>
+                  </table>
+                </div>
+
+
+         <div class="foot">
+             <p style="text-align: center;">Report Pulled by: {% if puller.email %}{{ puller.email }} {% else %} {{ puller.name }}{% endif %}</p>
+         </div>
+
+     </body>
+ </html>
