+++ templates/dashboard/reports/orders/charts/pdf/product_pdf.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title> {% load staticfiles %}
+ <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+ <html>
+    <link href='https://fonts.googleapis.com/css?family=Nunito:400,300' rel='stylesheet' type='text/css'>
+     <head>
+         <title>Item Report</title>
+         <style type="text/css">
+             body {
+                   font-family: 'Nunito', sans-serif;
+                   color: #384047;
+             }
+             .header {
+                 font-size: 15px;
+                 font-weight: 100;
+                 text-align: center;
+                 color: #007cae;
+             }
+             .title {
+                 font-size: 16px;
+                 font-weight: 100;
+                /* text-align: right;*/
+                padding: 10px 20px 0px 20px;
+             }
+             .title span {
+                 color: #007cae;
+             }
+             .detailss {
+                 padding: 10px 20px 0px 20px;
+                 text-align: left !important;
+                 /*margin-left: 40%;*/
+             }
+             .hrItem {
+                 border: none;
+                 height: 1px;
+                 /* Set the hr color */
+                 color: #333; /* old IE */
+                 background-color: #fff; /* Modern Browsers */
+             }
+             .img-circle{
+              border-radius: 50%;
+             }
+             table {
+              max-width: 960px;
+              margin: 10px auto;
+            }
+            th {
+                  font-weight: 400;
+                  background: #1779c6;
+                  color: #FFF;
+                }
+             .td{text-align: center;}
+             tr{padding:5px;}
+            .foot{padding-top:30px;}
+            .foot p{ color: #333;font-size: 12px;}
+            .chart{width:500px;height: 400px;}
+            .data-table-container{padding-top:20px;}
+         </style>
+     </head>
+     <body>
+         <div class='wrapper'>
+             <div class='header'>
+                 <p class='title'>Sales Summary Report <br>
+                   <span style="font-size: 0.8em;">{{today| date:'jS F Y'}}</span>
+                 </p>
+             </div>
+         <div>
+             <img alt="embedded" src="data:image/png;base64,{{image}}"/>
+
+
+         <div class="foot">
+             <p style="text-align: center;">Report Pulled by: {% if puller.email %}{{ puller.email }} {% else %} {{ puller.name }}{% endif %}</p>
+         </div>
+
+     </body>
+ </html></title>
+</head>
+<body>
+
+</body>
+</html>
