+++ templates/dashboard/reports/orders/paginate.html
@@ -1,0 +1,27 @@
+{% load staticfiles %}
+{% load prices_i18n %}
+{% load humanize %}
+ {% if sales %}
+    {% for sale in sales %}
+      <tr class='clickable-row' id="{{sale.pk }}">
+        <td class="td" data-href="{% url 'dashboard:sale-detail' pk=sale.pk %}">{{sale.created|date:"d-m-Y H:i a"}}</td>
+        <td class="td" data-href="{% url 'dashboard:sale-detail' pk=sale.pk %}">{{sale.invoice_number }}</td>
+          <td class="td" data-href="{% url 'dashboard:sale-detail' pk=sale.pk %}">{% if sale.customer_name %}{{sale.customer_name}}{% else %} Customer {% endif %}</td>
+        <td class="td" data-href="{% url 'dashboard:sale-detail' pk=sale.pk %}">{% if sale.user.name %}{{sale.user.name}}{% else %}{{sale.user.email}}{% endif %}</td>
+        <td class="td" data-href="{% url 'dashboard:sale-detail' pk=sale.pk %}">{% if sale.payment_options.name %}{{sale.payment_options.name}}{% else %} Cash {% endif %}</td>
+        <td class="td" data-href="{% url 'dashboard:sale-detail' pk=sale.pk %}">{{sale.quantity}}</td>
+        <td class="td" data-href="{% url 'dashboard:sale-detail' pk=sale.pk %}"><span class="pull-right">{{sale.total_net|intcomma }}</span></td>
+          <td class="td" data-href="{% url 'dashboard:sale_tax_detail' pk=sale.pk %}"><span class="pull-right">{{sale.total_tax|intcomma }}</span></td>
+      </tr>
+        <script type="text/javascript">
+        $(document).ready(function(){
+
+            $(".td").click(function(){
+                  window.location = $(this).data("href");
+              });
+        });
+        </script>
+    {% endfor %}
+    {% else %}
+     <tr><td colspan="7"><h2 class="text-center">no sales available</h2></td></tr>
+    {% endif %}
