+++ templates/dashboard/reports/orders/pdf/pdf.html
@@ -1,0 +1,88 @@
+{% load staticfiles %}
+ <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+ <html>
+     <head>
+         <title>Sales Items IReport</title>
+         <style type="text/css">
+             body {sans-serif; color: #384047;}
+             .header {font-family: 'Verdana';font-size: 15px;font-weight: 100;text-align: center;color: #007cae;}
+             .title {font-size: 16px;font-weight: 100;}
+             .header p{color:#384047;}
+
+             .img-circle{border-radius: 50%;}
+             table {ax-width: 960px; margin: 10px auto;}
+            tr{border-bottom:0.5px solid #ddd;border-right:0.5px solid #ddd;
+                border-left:0.5px solid #ddd;
+            }
+            td{border:none;}
+            th {
+                  font-weight: bold;
+                  background:#2196F3;
+                  color: #FFF;
+                  border:none;
+                  font-size:11px;
+                  vertical-align:middle;
+                }
+             .td{text-align: center;font-size:11px;}
+             tr{padding:5px;}
+             .img{width:50px;height:50px;}
+         </style>
+     </head>
+     <body>
+         <div class='wrapper'>
+             <div class='header'>
+                 <p class='title' style="margin-top:2px;margin-bottom:0px;font-size:13px;"> SALES ITEMS REPORT
+                 <p style="margin-bottom:0px;margin-top:2px;font-size:11px;">
+                        CUSTOMER: {% if sale.customer.name %} {{sale.customer.name|upper}} {% else %} -{% endif %}
+                 </p>
+                     <p style="margin-bottom:0px;margin-top:2px;font-size:11px;">
+                        RECEIPT: {{sale.invoice_number}}
+                 </p>
+                 <p style="margin-bottom:0px;margin-top:2px;font-size:10px;">
+                        SERVED BY: {{sale.user.name | upper}}
+                 </p>
+                    <p style="margin-top:2px;font-size:11px;">{{today | date:'jS F Y'}}</p>
+                 </p>
+             </div>
+         <div>
+         <div class='details'>
+            <table class="table  table-hover datatable-header-footer" style="border:0.5px solid #ddd;">
+                <thead>
+                  <tr class="bg-primary">
+                       <th>Transaction Date</th>
+                          <th>Item Name</th>
+                          <th>Unit Cost</th>
+                          <th>Quantity</th>
+                          <th>Total Sales</th>
+                  </tr>
+                </thead>
+                <tbody id="tb">
+                    {% if items %}
+                        {% for item in items %}
+                          <tr>
+                            <td>{{item.sales.created|date:"d-m-Y H:i a"}}</td>
+                            <td>{{item.product_name}}</td>
+                            <td>{{item.unit_cost}}</td>
+                            <td>{{item.quantity}}</td>
+                            <td>{{item.total_cost}}</td>
+                          </tr>
+                        {% endfor %}
+                         {% else %}
+                         <tr><td colspan="5"><h2 class="text-center">No Items Available</h2></td></tr>
+                        {% endif %}
+                  </tbody>
+           </table>
+         </div>
+         <div class="foot" id="footerContent">
+                 <p style="text-align: center;font-size:11px;">Report Pulled by:
+                     <span style="font-weight:bold">
+                         {% if puller.name %}
+                            {{ puller.name | upper }}
+                         {% else %}
+                            {{ puller.email | upper }}
+                         {% endif %}
+                     </span></p>
+
+         </div>
+     </body>
+ </html>
